chore: move message class
