+++ packages/crypto/__tests__/utils/message.test.js
@@ -1,4 +1,4 @@
-const { Message } = require('../../lib/utils')
+const { Message } = require('../../lib/crypto')
 
 const fixture = {
   data: {
