+++ packages/crypto/lib/crypto/index.js
@@ -1,6 +1,7 @@
 module.exports = {
   crypto: require('./crypto'),
   hdwallet: require('./hdwallet'),
+  Message: require('./message'),
   slots: require('./slots'),
   utils: require('./utils')
 }
