+++ packages/crypto/lib/utils/index.js
@@ -2,7 +2,6 @@
 
 module.exports = {
   Bignum: require('./bignum'),
-  Message: require('./message'),
   formatArktoshi: require('./format-arktoshi'),
   sortTransactions: require('./sort-transactions')
 }
